+++ js/server.js
@@ -80,7 +80,7 @@ function Server(config) {
 			app.use("/js", express.static(__dirname));
 
 			// TODO add tests directory only when running tests?
-			const directories = ["/config", "/css", "/fonts", "/modules", "/vendor", "/translations", "/tests/configs", "/tests/mocks"];
+			const directories = ["/config", "/css", "/fonts", "/modules", "/vendor", "/translations", "/layout", "/tests/configs", "/tests/mocks"];
 			for (const directory of directories) {
 				app.use(directory, express.static(path.resolve(global.root_path + directory)));
 			}
@@ -89,7 +89,7 @@ function Server(config) {
 
 			app.get("/version", (req, res) => getVersion(req, res));
 
-			app.get("/config", (req, res) => getConfig(req, res));
+			app.get("/config/:configFile", (req, res) => getConfig(req, res));
 
 			app.get("/", (req, res) => getHtml(req, res));